2 hull sigmoid working
Added a sigmix function that takes two distances and returns the
amount to subtract from 1.0 to find the proportion to allocate to
the first hull, with the remainder going to the other hull.

Applied this in the predictor so that at least for two classes it
seems to be mixing as per the paper. Still need to handle the 3 and
4 cases, which are an extension of this.