Further sigmoid handling improvements.
Improved it a bit more, but realized there seems to be some problem
with the distance computations. I'm not sure the scipy spatial
functions are working the way I thought they were, so I'm investigating
that now. This commit adds a gradient based test page that helped
identify the issue, and a script to simplify what I'm giving to scipy
to help further target the issue. With the geometry I created, going
up one axis should yield easily computable distances that are exactly
the distance from the face as I put on the L* axis, but its not working
that way. I've been reviewing the scipy and Qhull documentation and
source but haven't been able to figure it out quite yet. More to do on
it, but this is a good checkpoint.