Fixed distance function; refactored hull class.
I learned that the Delaunay code and and plane distance function
wasn't working the way I expected. I refactored the hull code a bit
so that Hull is a new class with some of the logic. I also found a
pointTriangleDistance function that does seem to work as I need and
plugged it in. The classifications for points outside the hulls are
much more believable now, but the code is noticeably slower. I think
there are several optimization opportunities, but as it does seem to
work quite a bit better now this seems to be a good checkpoint.