Refactor to be more packaging friendly.
The models and associated diagnostics and helper functions went
into a new colornames package, which perhaps we can eventually
make available via pip for easier installation. The scripts went
into a separate Scripts folder, and could be installed in the
main Python Scripts folder during installation. Still need to
refactor some things within the package and scripts, but this
should be enough to make it a bit cleaner to start implementing
different models using the same scripts.