color-names-9 selection could exclude more on outer edge
Added another constraint to the selection: if the color is within
only one hull, then in addition to being close to its hull's boundary
it must also be within 10 Delta-E of the next closest hull's boundary.
This helps address the problem of points on the "back" of the hull.

Also added line num to the html diagnostic file.